+++ src/components/Header.js
@@ -2,7 +2,7 @@ function Header() {
   return (
     <div className="mb-12">
       <h1 className="mb-1 title-h1 text-orange">
-        Word Limit
+        Count My Words
       </h1>
       <h3>
         A simple word and character counter. Free to use.